New Video Player Updates
Started removing the default double click behaviour of the video.
When dbl clicking the video, it now increases the video size itself, making it full screen.
Added the new scrolling lyrics div as well as a button to show or hide them.

Still to do:
The lyrics should start scrolling from the lower portion of the screen, taking into consideration the videos scroll bar and still available buttons.